services/transcription-service/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pydantic-settings
python-dotenv
redis[hiredis]
celery
requests

pyannote.audio==3.3.2
numpy==1.26.4
soundfile==0.12.1
httpx==0.27.0
huggingface_hub==0.19.4